Moved DataSourceSelect into related folder. Changed SelectedItem name. Updated DataSourceSelect imports.
